Add switch in CMake to choose the right nvjpegDecoder w.r.t. the nvjpeg API
